Peoject/project2/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.3
matplotlib>=3.7.1
Pillow>=10.0.0    